+++ meta_package_manager/managers/snap.py
@@ -39,7 +39,7 @@ class Snap(PackageManager):
 
         .. code-block:: shell-session
 
-            $ snap --version
+            ► snap --version
             snap       2.44.1
             snapd      2.44.1
             series     16
@@ -57,7 +57,7 @@ class Snap(PackageManager):
         Raw CLI output samples:
 
         .. code-block:: shell-session
-            $ snap list
+            ► snap list
             Name    Version    Rev   Aufzeichnung   Herausgeber     Hinweise
             core    16-2.44.1  8935  latest/stable  canonical✓      core
             wechat  2.0        7     latest/stable  ubuntu-dawndiy  -
@@ -82,7 +82,7 @@ class Snap(PackageManager):
         """ Fetch matching packages from ``snap find`` output.
 
         .. code-block:: shell-session
-            $ snap find doc
+            ► snap find doc
             Name       Version      Herausgeber  Hinweise  Zusammenfassung
             journey    2.14.3       2appstudio   -         Your private diary.
             nextcloud  17.0.5snap1  nextcloud✓   -         Nextcloud Server
@@ -128,7 +128,7 @@ class Snap(PackageManager):
 
         .. code-block:: shell-session
 
-            $ snap refresh --list
+            ► snap refresh --list
             Name            Version  Rev  Herausgeber     Hinweise
             standard-notes  3.3.5    8    standardnotes✓  -
         """