+++ CHANGES.rst
@@ -18,6 +18,7 @@ Changelog
 * [bitbar] Test plugin with Python 2.7.
 * [mpm] Use local copy of ``boltons`` utils while we wait for upstream release.
 * [mpm] Allow for manager-specific search path to help hunting down CLIs.
+* [mpm] Highlight CLI and indent results in debug output.
 * [mpm] Bump dependency to ``pylint 2.5`` and ``cli-helpers 2.0``.
 * [mpm] Use local copy of ``boltons`` utils while we wait for upstream release.
 * [mpm] Move pylint config from ``setup.cfg`` to ``pyproject.toml``.